+++ src/pages/Private/LabelList/index.styled.ts
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
-import { AddLabelField } from '@/components/Molecules/AddLabelField/index.styled';
+import { LabelEditForm } from '@/components/Molecules/LabelEditForm/index.styled';
 import { LabelTable } from '@/components/Skeleton/LabelTable/index.styled';
 import { StyledNavLink } from '@/components/Molecules/NavLink/index.styles';
 
 export const LabelList = styled.div`
-  & > ${AddLabelField} {
+  & > ${LabelEditForm} {
     margin-bottom: 24px;
     border: 1px solid ${({ theme }) => theme.COLORS.LINE};
   }
